+++ modules/front-end/src/app/core/services/project.service.ts
@@ -5,6 +5,8 @@ import { environment } from 'src/environments/environment';
 import { IProject, IProjectEnv } from '@shared/types';
 import { CURRENT_PROJECT } from "@utils/localstorage-keys";
 import { MessageQueueService } from "@services/message-queue.service";
+import {PermissionsService} from "@services/permissions.service";
+import {permissionActions} from "@shared/policy";
 
 @Injectable({
   providedIn: 'root'
@@ -14,15 +16,24 @@ export class ProjectService {
 
   constructor(
     private http: HttpClient,
+    private permissionsService: PermissionsService,
     private messageQueueService: MessageQueueService
   ) { }
 
-  getList(): Observable<IProject[]> {
-    return this.http.get<IProject[]>(this.baseUrl);
-  }
-
   async getListAsync(): Promise<IProject[]> {
-    return firstValueFrom(this.getList());
+    const projects = await firstValueFrom(this.http.get<IProject[]>(this.baseUrl));
+
+    return projects.filter((project) => {
+      const rn = this.permissionsService.getProjectRN(project.name);
+      return this.permissionsService.isGranted(rn, permissionActions.ListProjects)
+    }).map((project) => {
+      project.environments = project.environments.filter((env) => {
+        const envRN = this.permissionsService.getEnvRN(project.name, env.name);
+        return !this.permissionsService.isDenied(envRN, permissionActions.AccessEnvs);
+      });
+
+      return project;
+    }).filter((project) => project.environments.length);
   }
 
   get(projectId: string): Observable<IProject> {
